Fix bug for qca4531

+++ qca4531/openwrt/4531_dailybuild.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -ex
 
 VER_PREFIX="4531"
 
@@ -89,23 +89,24 @@ function add_version()
     sed -i "${LINE_NUM}iEXTRAVERSION = ${DAILYBUILD_VER}" $ROOT_DIR/$U_BOOT_PATH
 
     # Set Linux version
-    rm $ROOT_DIR/$KERNEL_PATH
-    echo "${DAILYBUILD_VER}" > $ROOT_DIR/$KERNEL_PATH
+    if [ -e $ROOT_DIR/$KERNEL_PATH ] ; then
+        rm $ROOT_DIR/$KERNEL_PATH
+    fi
+    echo "\"${DAILYBUILD_VER}\"" > $ROOT_DIR/$KERNEL_PATH
 }
 
 function build_images()
 {
+    add_version
     cd $CURR_PATH/$ROOT_DIR/qsdk
 
     make package/symlinks
     cp qca/configs/advantech/${NEW_MACHINE}.config .config
     make defconfig
-
-    add_version
     make V=s
 
     if [ "$?" -ne 0 ]; then
-        echo "[ADV] Build failure! Check details in ${LOG_DIR}.tgz"
+        echo "[ADV] Build failure! Check details in Jenkins console output."
         exit 1
     fi
 }
